Moved Files of CreatePayIndiscord
