--- src/main/java/space/space_spring/domain/discord/application/port/in/CreatePayInDiscordUseCase.java
@@ -1,6 +1,0 @@
-package space.space_spring.domain.discord.application.port.in;
-
-public interface CreatePayInDiscordUseCase {
-
-    Long createPayInDiscord(CreatePayInDiscordCommand command);
-}
