+++ src/main/java/space/space_spring/domain/discord/application/port/in/createPay/CreatePayInDiscordCommand.java
@@ -1,4 +1,4 @@
-package space.space_spring.domain.discord.application.port.in;
+package space.space_spring.domain.discord.application.port.in.createPay;
 
 import lombok.Builder;
 import lombok.Getter;
