+++ src/main/java/space/space_spring/domain/discord/application/port/in/createPay/TargetOfCreatePayInDiscordCommand.java
@@ -1,4 +1,4 @@
-package space.space_spring.domain.discord.application.port.in;
+package space.space_spring.domain.discord.application.port.in.createPay;
 
 import lombok.Getter;
 import space.space_spring.domain.pay.domain.Money;
