+++ src/main/java/space/space_spring/domain/pay/application/service/CreatePayService.java
@@ -3,9 +3,9 @@ package space.space_spring.domain.pay.application.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import space.space_spring.domain.discord.application.port.in.CreatePayInDiscordUseCase;
-import space.space_spring.domain.discord.application.port.in.CreatePayInDiscordCommand;
-import space.space_spring.domain.discord.application.port.in.TargetOfCreatePayInDiscordCommand;
+import space.space_spring.domain.discord.application.port.in.createPay.CreatePayInDiscordUseCase;
+import space.space_spring.domain.discord.application.port.in.createPay.CreatePayInDiscordCommand;
+import space.space_spring.domain.discord.application.port.in.createPay.TargetOfCreatePayInDiscordCommand;
 import space.space_spring.domain.pay.application.port.in.createPay.CreatePayCommand;
 import space.space_spring.domain.pay.application.port.in.createPay.CreatePayUseCase;
 import space.space_spring.domain.pay.application.port.in.createPay.TargetOfCreatePayCommand;
