+++ src/test/java/space/space_spring/domain/pay/application/service/CreatePayServiceTest.java
@@ -5,8 +5,8 @@
 //import org.junit.jupiter.api.DisplayName;
 //import org.junit.jupiter.api.Test;
 //import org.mockito.Mockito;
-//import space.space_spring.domain.discord.application.port.in.CreatePayInDiscordUseCase;
-//import space.space_spring.domain.discord.application.port.in.CreatePayInDiscordCommand;
+//import space.space_spring.domain.discord.application.port.in.createPay.CreatePayInDiscordUseCase;
+//import space.space_spring.domain.discord.application.port.in.createPay.CreatePayInDiscordCommand;
 //import space.space_spring.domain.pay.adapter.in.web.createPay.TargetOfPayRequest;
 //import space.space_spring.domain.pay.application.port.in.createPay.CreatePayCommand;
 //import space.space_spring.domain.pay.application.port.out.CreatePayPort;
